add batch and stream operators
