+++ pixels-common/src/main/resources/pixels.properties
@@ -178,6 +178,8 @@ executor.selectivity.enabled=true
 executor.intra.worker.parallelism=8
 # which cloud function service to use, can be lambda (AWS Lambda) or vhive (vHive)
 executor.function.service=lambda
+# which method will be used for data exchanging, batch or stream
+executor.exchange.method=batch
 executor.ordered.layout.enabled=false
 executor.compact.layout.enabled=true
 # the names of the cloud function workers
