+++ pixels-planner/src/main/java/io/pixelsdb/pixels/planner/plan/physical/Operator.java
@@ -21,16 +21,14 @@ package io.pixelsdb.pixels.planner.plan.physical;
 
 import io.pixelsdb.pixels.common.utils.ConfigFactory;
 
-import java.util.concurrent.*;
-
-import static com.google.common.base.Preconditions.checkArgument;
-import static java.util.Objects.requireNonNull;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
 /**
  * @author hank
  * @create 2022-07-05
  */
-public abstract class Operator
+public abstract class Operator implements OperatorExecutor
 {
     protected static final double StageCompletionRatio;
     protected static final ExecutorService operatorService = Executors.newCachedThreadPool();
@@ -56,86 +54,5 @@ public abstract class Operator
     public String getName()
     {
         return name;
-    }
-
-    /**
-     * Execute this operator recursively.
-     *
-     * @return the completable future that completes when this operator is complete, and
-     * provides the computable futures of the outputs of this operator
-     */
-    public abstract CompletableFuture<CompletableFuture<?>[]> execute();
-
-    /**
-     * Execute the previous stages (if any) before the last stage, recursively.
-     * And return the completable future that completes when the previous states are complete.
-     *
-     * @return empty array if the previous stages do not exist or do not need to be waited for
-     */
-    public abstract CompletableFuture<Void> executePrev();
-
-    /**
-     * This method collects the outputs of the operator. It may block until the join
-     * completes, therefore it should not be called in the query execution thread. Otherwise,
-     * it will block the query execution.
-     * @return the outputs of the workers in this operator
-     */
-    public abstract OutputCollection collectOutputs() throws ExecutionException, InterruptedException;
-
-    public interface OutputCollection
-    {
-        long getTotalGBMs();
-
-        int getTotalNumReadRequests();
-
-        int getTotalNumWriteRequests();
-
-        long getTotalReadBytes();
-
-        long getTotalWriteBytes();
-
-        long getLayerInputCostMs();
-
-        long getLayerComputeCostMs();
-
-        long getLayerOutputCostMs();
-    }
-
-    public static void waitForCompletion(CompletableFuture<?>[] stageOutputs) throws InterruptedException
-    {
-        waitForCompletion(stageOutputs, StageCompletionRatio);
-    }
-
-    public static void waitForCompletion(CompletableFuture<?>[] stageOutputs, double completionRatio)
-            throws InterruptedException
-    {
-        requireNonNull(stageOutputs, "stageOutputs is null");
-
-        if (stageOutputs.length == 0)
-        {
-            return;
-        }
-
-        while (true)
-        {
-            double completed = 0;
-            for (CompletableFuture<?> childOutput : stageOutputs)
-            {
-                if (childOutput.isDone())
-                {
-                    checkArgument(!childOutput.isCompletedExceptionally(),
-                            "worker in the stage is completed exceptionally");
-                    checkArgument(!childOutput.isCancelled(),
-                            "worker in the stage is cancelled");
-                    completed++;
-                }
-            }
-
-            if (completed / stageOutputs.length >= completionRatio)
-            {
-                break;
-            }
-            TimeUnit.MILLISECONDS.sleep(10);
-        }
     }
 }
